16/03/21 - Grafica histograma - grises - negativo - Binario (Terminar)

+++ src/espacial/Histograma.java
@@ -16,15 +16,15 @@ import java.awt.image.BufferedImage;
 public class Histograma {
     
     private Image imagenOriginal;
-    private int[]r;
-    private int[]g;
-    private int[]b;
+    private double[]r;
+    private double[]g;
+    private double[]b;
 
     public Histograma(Image imagen) {
         this.imagenOriginal = imagen;
-        this.r = new int[256];
-        this.g = new int[256];
-        this.b = new int[256];
+        this.r = new double[256];
+        this.g = new double[256];
+        this.b = new double[256];
     }
     public void calcularHistogramas(){
         // recorrer la imagen
@@ -53,21 +53,21 @@ public class Histograma {
     /**
      * @return the r
      */
-    public int[] getR() {
+    public double[] getR() {
         return r;
     }
 
     /**
      * @return the g
      */
-    public int[] getG() {
+    public double[] getG() {
         return g;
     }
 
     /**
      * @return the b
      */
-    public int[] getB() {
+    public double[] getB() {
         return b;
     }
